Consolidate declaration merging
